refactor: refactor filenames for better folder structuring
