+++ src/components/index.ts
@@ -1,5 +1,5 @@
-import Card from './card';
-import EmptyMessage from './empty-message';
+import Card from './card/Card';
+import EmptyMessage from './empty-message/EmptyMessage';
 import OrganizationList from './organization/OrganizationList';
 import SearchResultsContainer from './search-results/Container';
 import SearchResultsHeader from './search-results/Header';
